+++ inst/@mp/mrdivide.m
@@ -2,5 +2,5 @@
 
 function result = mrdivide (varargin)
   error ("mplapack:NotImplemented", ...
-         "mp right division is not implemented before M09");
+         "mp right division is not implemented yet");
 endfunction